requirements.txt
alembic==1.18.4
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.2.0
attrs==25.4.0
babel==2.18.0
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
cachetools==5.5.2
catboost==1.2.10
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.2
cryptography==46.0.5
cycler==0.12.1
databricks-sdk==0.91.0
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.8.0
docker==7.1.0
email-validator==2.3.0
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.131.0
fastapi-cli==0.0.23
fastapi-cloud-cli==0.13.0
fastar==0.8.0
fastjsonschema==2.21.2
Flask==3.1.3
flask-cors==6.0.2
fonttools==4.61.1
fqdn==1.5.1
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.48.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
graphviz==0.21
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huey==2.6.0
idna==3.11
imbalanced-learn @ file:///C:/Users/task_176795969927117/croot/imbalanced-learn_1767959802587/work
importlib_metadata==7.2.1
ipykernel==7.2.0
ipython==8.38.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib @ file:///C:/Users/task_176786734930694/croot/joblib_1767868144277/work
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
lark==1.3.1
lightgbm==4.6.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
mkl-service==2.5.2
mkl_fft @ file:///C:/miniconda3/conda-bld/mkl_fft_1761592981169/work
mkl_random @ file:///C:/miniconda3/conda-bld/mkl_random_1761593075610/work
mlflow==3.10.0
mlflow-skinny==3.10.0
mlflow-tracing==3.10.0
narwhals==2.16.0
nbclient==0.10.4
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.5.3
notebook_shim==0.2.4
numpy==1.26.4
opentelemetry-api==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
overrides==7.7.0
packaging==23.2
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.6
pillow==10.4.0
platformdirs==4.9.4
plotly==6.6.0
prettytable==3.17.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
protobuf==4.25.8
psutil==7.2.2
pure_eval==0.2.3
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==4.0.0
pytz==2026.1.post1
pywin32==311
pywinpty==3.0.3
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==13.9.4
rich-toolkit==0.19.7
rignore==0.7.6
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.3.2
scipy @ file:///C:/Users/task_177627752328296/croot/scipy_1776277626143/work/dist/scipy-1.15.3-cp310-cp310-win_amd64.whl#sha256=4b31ae66cb62ee7d0d6cb492e8af90b64e32541a8bee52f4e127b96528004419
seaborn==0.13.2
Send2Trash==2.1.0
sentry-sdk==2.54.0
shellingham==1.5.4
six==1.17.0
sklearn-compat @ file:///C:/Users/task_176789484799927/croot/sklearn-compat_1767894975643/work
skops==0.13.0
smmap==5.0.3
soupsieve==2.8.3
SQLAlchemy==2.0.48
sqlparse==0.5.5
stack-data==0.6.3
starlette==0.52.1
streamlit==1.31.1
tenacity==8.5.0
terminado==0.18.1
threadpoolctl @ file:///C:/miniconda3/conda-bld/threadpoolctl_1764361785970/work
tinycss2==1.5.1
toml==0.10.2
tomli==2.4.0
tornado==6.5.5
traitlets==5.14.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.6.3
uvicorn==0.41.0
validators==0.35.0
waitress==3.0.2
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.6
widgetsnbextension==4.0.15
xgboost==3.2.0
zipp==3.23.0
